requirements.txt
pillow
torchvision
torch
pandas
